+++ mlquantify/elm/__init__.py
@@ -1,0 +1,19 @@
+from ._svmperf import MultivariateLossSVM
+from ._classes import (
+    ELM,
+    SVMQ,
+    SVMKLD,
+    SVMNKLD,
+    SVMAE,
+    SVMRAE,
+)
+
+__all__ = [
+    "MultivariateLossSVM",
+    "ELM",
+    "SVMQ",
+    "SVMKLD",
+    "SVMNKLD",
+    "SVMAE",
+    "SVMRAE",
+]